Merging dev at daa52dd, which is the base of the OCD georeferencing
PR #1034 at the moment.
Updating for new Symbol duplication API.

+++ src/core/renderables/renderable_implementation.cpp
@@ -3,7 +3,7 @@
  *    Copyright 2012-2017 Kai Pastor
  *
  *    This file is part of OpenOrienteering.
- *
+ *	
  *    OpenOrienteering is free software: you can redistribute it and/or modify
  *    it under the terms of the GNU General Public License as published by
  *    the Free Software Foundation, either version 3 of the License, or
@@ -725,7 +725,7 @@ TextRenderable::TextRenderable(const TextSymbol* symbol, const TextObject* text_
 	QTransform t { 1.0, 0.0, 0.0, 1.0, anchor_x, anchor_y };
 	t.scale(scale_factor, scale_factor);
 	
-	auto rotation_rad = qreal(text_object->getRotation());
+	auto rotation_rad = text_object->getRotation();
 	if (!qIsNull(rotation_rad))
 	{
 		rotation = -qRadiansToDegrees(rotation_rad);
